update on the advance problem solution

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,8 @@
-income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 monthly_savings = income-monthly_expenses
-project_savings = int(monthly_savings * 12 + (monthly_savings*12*0.05))
+project_savings = float(monthly_savings * 12 + (monthly_savings*12*0.05))
 print("Your monthly savings are $"+str(monthly_savings))
 print("Projected Savings after one year, with interest, is: $" + str(project_savings))
+
+
